requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy==1.17.2
torch==1.10.1+cpu
torchvision==0.11.2+cpu

scipy==1.4.1
scikit-image==0.16.2
Pillow==7.1.2
